Fixed server startup issues
