+++ enderio-armory/src/main/java/com/enderio/armory/common/item/darksteel/upgrades/nightvision/NightVisionHandler.java
@@ -1,6 +1,5 @@
 package com.enderio.armory.common.item.darksteel.upgrades.nightvision;
 
-import com.enderio.armory.client.KeyBinds;
 import com.enderio.armory.common.capability.DarkSteelHelper;
 import com.enderio.armory.common.config.ArmoryConfig;
 import com.enderio.armory.common.init.ArmoryDataComponents;
@@ -10,16 +9,12 @@ import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.neoforged.neoforge.client.event.ClientTickEvent;
 import net.neoforged.neoforge.event.tick.PlayerTickEvent;
-import net.neoforged.neoforge.network.PacketDistributor;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
 public class NightVisionHandler {
@@ -52,22 +47,6 @@ public class NightVisionHandler {
         appliedPlayers.add(player.getUUID());
     }
 
-    public void toggleNightVision(ClientTickEvent.Post event) {
-        LocalPlayer player = Minecraft.getInstance().player;
-        if (player == null) {
-            return;
-        }
-        Optional<ItemStack> eq = getEquippedHelmet(player);
-        if (eq.isEmpty()) {
-            return;
-        }
-        while (KeyBinds.NIGHT_VISION_MAPPING.get().consumeClick()) {
-            boolean newVal = !eq.get().getOrDefault(ArmoryDataComponents.DARK_STEEL_NIGHT_VISION_ACTIVE, false);
-            eq.get().set(ArmoryDataComponents.DARK_STEEL_NIGHT_VISION_ACTIVE, newVal);
-            PacketDistributor.sendToServer(new NightVisionEnabledPacket(newVal));
-        }
-    }
-
     public void handleActivePacket(NightVisionEnabledPacket packet, IPayloadContext context) {
         context.enqueueWork(() -> {
             var player = context.player();
@@ -87,7 +66,7 @@ public class NightVisionHandler {
         appliedPlayers.remove(player.getUUID());
     }
 
-    private Optional<ItemStack> getEquippedHelmet(Player player) {
+    public static Optional<ItemStack> getEquippedHelmet(Player player) {
         ItemStack equipped = player.getItemBySlot(EquipmentSlot.HEAD);
         if (equipped.is(ArmoryTags.Items.DARK_STEEL_UPGRADEABLE_HELMET)
                 && (DarkSteelHelper.hasUpgrade(equipped, NightVisisionUpgrade.NAME))) {
